+++ Assets/Script/Song/Exporters/WebBrowserDifficulty.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1dc94d6814be43ec95a4f4a10ebc31e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
