+++ Assets/Script/Song/Exporters/WebBrowserExport.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 245483ffb8bc450ba30cedde92c28bc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
