+++ Assets/Script/Song/Exporters/WebBrowserGenreTable.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace YARG.Song
+namespace YARG.Song.Exporters
 {
     /// <summary>
     /// Genre interner for compact web export.
