+++ Assets/Script/Song/Exporters/WebBrowserGenreTable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f589f75c487e4ebfb3071de401016403
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
