+++ Assets/Script/Song/Exporters/WebBrowserRecord.cs
@@ -6,7 +6,7 @@ using YARG.Core;
 using YARG.Core.Song;
 using YARG.Core.Utility;
 
-namespace YARG.Song
+namespace YARG.Song.Exporters
 {
     /// <summary>
     /// Compact song record for web browser export.
