+++ Assets/Script/Song/Exporters/WebBrowserRecord.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85c328033ebc474baa06ac4a6734201b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
