+++ Assets/Script/Song/Exporters/WebBrowserTemplate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4d4647993834b50b43847d3d7ace255
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
